requirements.txt
streamlit
pandas
numpy
yfinance
requests
pytz
python-dotenv
schwabdev
supabase>=2.6.0
python-dateutil
lxml